Make Google Maps search location aware.
Previously, ambiguous addresses like 'Union Street' would just return the best
match across the US which is not necessarily what the user wants.
With this commit, for ambiguous addresses we use users current location
or their device location as a signal to bias Google Maps to return
results closer to the user. We do this only for ambiguous addresses but when the
provides the city or state or zipcode, we just do a simple search.